BUG FIX: Starting in commit 1334060, only one file for each directory was stored in the cache. This has been fixed.
When a directory is removed, all files that are within that directory are removed from the cache.

Editing a directory will no longer re-initialize the cache. Rather, it will remove the files that we
re present in the old directory, and add the files that are in the new directory.